@@ -30,18 +30,18 @@
   },
   "devDependencies": {
     "@sveltejs/adapter-auto": "^3.0.0",
-    "@sveltejs/kit": "^2.0.0",
-    "@sveltejs/package": "^2.2.1",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@sveltejs/kit": "^2.5.27",
+    "@sveltejs/package": "^2.3.7",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@testing-library/svelte": "^4.0.3",
     "rome": "12.1.3",
     "jsdom": "^22.1.0",
     "publint": "^0.2.3",
-    "svelte": "^4.2.7",
-    "svelte-check": "^3.5.2",
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.0.0",
     "tslib": "^2.6.1",
-    "typescript": "^5.1.6",
-    "vite": "^5.0.0",
+    "typescript": "^5.5.0",
+    "vite": "^5.4.4",
     "vitest": "^1.0.0"
   },
   "svelte": "./dist/index.js",
